7 enemy actions (#27)
* enemy spawning from level_info

* enemy movement start

* clippy and small bug

* enemy firing sprite swap and event

* remove old comment

* fix input text location

* resolve comments

* flip sprites, add random shot chance

* remove print

+++ src/game/enemy/events.rs
@@ -1,4 +1,4 @@
 use bevy::prelude::Event;
 
 #[derive(Event)]
-pub struct EnemyShotEvent(pub usize);
+pub struct EnemyShotEvent(pub u32);
